requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
opencv-python
joblib

openai == 2.15.0
requests == 2.32.5
slack_sdk == 3.39.0
python-dotenv

# Google Drive API (for face image download & cache)
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.100.0

# For production OpenFace integration
# opencv-python>=4.8.0
